Feature/login

See merge request cloud_track/class_01/web_project3/team02/santa__back!56

+++ src/main/java/com/example/santa/domain/meeting/service/MeetingService.java
@@ -13,21 +13,22 @@ import com.example.santa.domain.meeting.repository.MeetingRepository;
 import com.example.santa.domain.meeting.repository.MeetingTagRepository;
 import com.example.santa.domain.meeting.repository.ParticipantRepository;
 import com.example.santa.domain.meeting.repository.TagRepository;
+import com.example.santa.domain.user.entity.Role;
 import com.example.santa.domain.user.entity.User;
 import com.example.santa.domain.user.repository.UserRepository;
 import com.example.santa.global.exception.ExceptionCode;
 import com.example.santa.global.exception.ServiceLogicException;
+import com.example.santa.global.util.S3ImageService;
+import com.example.santa.global.util.mapsturct.ParticipantsDtoMapper;
 import jakarta.transaction.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -39,37 +40,59 @@ public class MeetingService {
     private final TagRepository tagRepository;
     private final MeetingTagRepository meetingTagRepository;
     private final ParticipantRepository participantRepository;
+    private final ParticipantsDtoMapper participantsDtoMapper;
+    private final S3ImageService s3ImageService;
 
-    public MeetingService(MeetingRepository meetingRepository, UserRepository userRepository, CategoryRepository categoryRepository, TagRepository tagRepository, MeetingTagRepository meetingTagRepository, ParticipantRepository participantRepository) {
+    public MeetingService(MeetingRepository meetingRepository, UserRepository userRepository, CategoryRepository categoryRepository, TagRepository tagRepository, MeetingTagRepository meetingTagRepository, ParticipantRepository participantRepository, ParticipantsDtoMapper participantsDtoMapper, S3ImageService s3ImageService) {
         this.meetingRepository = meetingRepository;
         this.userRepository = userRepository;
         this.categoryRepository = categoryRepository;
         this.tagRepository = tagRepository;
         this.meetingTagRepository = meetingTagRepository;
         this.participantRepository = participantRepository;
+        this.participantsDtoMapper = participantsDtoMapper;
+        this.s3ImageService = s3ImageService;
     }
 
     public MeetingResponseDto createMeeting(MeetingDto meetingDto){
+        // dto에서 불러온 카테고리명으로 카테고리를 가져옴
         Category category = categoryRepository.findByName(meetingDto.getCategoryName())
                 .orElseThrow(() -> new ServiceLogicException(ExceptionCode.CATEGORY_NOT_FOUND));
+        // 현재 로그인 한 유저를 불러옴
         User leader = userRepository.findByEmail(meetingDto.getUserEmail())
                 .orElseThrow(() -> new ServiceLogicException(ExceptionCode.USER_NOT_FOUND));
+
+        // 이미 같은 날짜에 다른 모임에 참여 중인지 확인
+        boolean isParticipatingOnSameDate = userRepository.findMeetingsByUserId(leader.getId()).stream()
+                .anyMatch(m -> m.getDate().equals(meetingDto.getDate()));
+
+        if (isParticipatingOnSameDate) {
+            // 같은 날짜에 다른 모임에 이미 참여중인 경우 예외 발생
+            throw new ServiceLogicException(ExceptionCode.ALREADY_PARTICIPATING_ON_DATE);
+        }
+
+        MultipartFile imageFile = meetingDto.getImageFile();
+        String imageUrl = "defaultUrl";
+        if (imageFile != null && !imageFile.isEmpty()) {
+            imageUrl = s3ImageService.upload(imageFile);
+        }
 
 
         Meeting meeting = Meeting.builder()
                 .meetingName(meetingDto.getMeetingName())
+                .leader(leader)
                 .category(category)
                 .mountainName(meetingDto.getMountainName())
                 .description(meetingDto.getDescription())
                 .headcount(meetingDto.getHeadcount())
                 .date(meetingDto.getDate())
-                .image(meetingDto.getImage())
+                .image(imageUrl)
                 .build();
 
         meetingRepository.save(meeting);
 
         Set<MeetingTag> meetingTags = new HashSet<>();
-
+        // dto에 있는 해시태그에서 생성되지 않은 해시태그면 생성해주고 생성되어 있으면 가져옴
         for (String tagName : meetingDto.getTags()) {
             Tag tag = tagRepository.findByName(tagName)
                     .orElseGet(() -> tagRepository.save(Tag.builder()
@@ -85,7 +108,7 @@ public class MeetingService {
         meeting.setMeetingTags(meetingTags);
 
         meetingRepository.save(meeting);
-
+        // 모임장을 참가자 목록에 추가해줌
         Participant participant = Participant.builder()
                 .user(leader)
                 .meeting(meeting)
@@ -99,7 +122,6 @@ public class MeetingService {
         return convertToDto(meetingRepository.save(meeting));
 
     }
-
     public MeetingResponseDto meetingDetail(Long id){
         Meeting meeting = meetingRepository.findById(id)
                 .orElseThrow(() -> new ServiceLogicException(ExceptionCode.MEETING_NOT_FOUND));
@@ -195,9 +217,17 @@ public class MeetingService {
         return convertToDto(meetingRepository.save(meeting));
     }
 
-    public void deleteMeeting(Long id) {
+    public void deleteMeeting(String email, Long id) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new ServiceLogicException(ExceptionCode.USER_NOT_FOUND));
+        Meeting meeting = meetingRepository.findById(id)
+                .orElseThrow(() -> new ServiceLogicException(ExceptionCode.MEETING_NOT_FOUND));
+
         if (!meetingRepository.existsById(id)) {
             throw new ServiceLogicException(ExceptionCode.MEETING_NOT_FOUND);
+        }
+        if (!Objects.equals(user.getId(), meeting.getLeader().getId()) || user.getRole() == Role.ADMIN){
+            throw new ServiceLogicException(ExceptionCode.USER_NOT_LEADER);
         }
         meetingRepository.deleteById(id);
     }
@@ -272,6 +302,20 @@ public class MeetingService {
             meetings = meetingRepository.findMeetingsByParticipantUserIdAndIdLessThan(user.getId(),lastId, PageRequest.of(0, size, Sort.by("id").descending()));
         }
         return meetings.map(this::convertToDto);
+    }
+
+    public List<ParticipantDto> endMeeting(String email, Long id) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new ServiceLogicException(ExceptionCode.USER_NOT_FOUND));
+        Meeting meeting = meetingRepository.findById(id)
+                .orElseThrow(() -> new ServiceLogicException(ExceptionCode.MEETING_NOT_FOUND));
+
+//        meeting.setEnd(true);
+//
+//        meetingRepository.save(meeting);
+
+        return participantsDtoMapper.toDtoList(meeting.getParticipant());
+
     }
 
     public MeetingResponseDto convertToDto(Meeting meeting) {
